renaming + add exports
