+++ packages/core-data/src/components/HitsPerPage.js
