+++ packages/core-data/src/index.js
@@ -14,12 +14,14 @@ export { default as FacetSlider } from './components/FacetSlider';
 export { default as FacetStateContextProvider } from './components/FacetStateContextProvider';
 export { default as FacetTimeline } from './components/FacetTimeline';
 export { default as HeaderImage } from './components/HeaderImage';
+export { default as HitsPerPage } from './components/HitsPerPage';
 export { default as KeyValueList } from './components/KeyValueList';
 export { default as LayerMenu } from './components/LayerMenu';
 export { default as LoadAnimation } from './components/LoadAnimation';
 export { default as MediaGallery } from './components/MediaGallery';
 export { default as Modal } from './components/Modal';
 export { default as OverlayLayers } from './components/OverlayLayers';
+export { default as Pagination } from './components/Pagination';
 export { default as PersistentSearchStateContextProvider } from './components/PersistentSearchStateContextProvider';
 export { default as PlaceDetails } from './components/PlaceDetails';
 export { default as PlaceLayersSelector } from './components/PlaceLayersSelector';
@@ -39,6 +41,7 @@ export { default as RelatedPlacesLayer } from './components/RelatedPlacesLayer';
 export { default as RelatedSources } from './components/RelatedSources';
 export { default as RelatedTaxonomies } from './components/RelatedTaxonomies';
 export { default as SearchResultsLayer } from './components/SearchResultsLayer';
+export { default as SearchResultsTable } from './components/SearchResultsTable';
 
 // Contexts
 export { default as CoreDataContext } from './context/CoreData';
